demux API を整理して fMP4 file demux を incremental 化する
