+++ src/lib.rs
@@ -12,8 +12,9 @@ mod boxes_moov_tree;
 mod boxes_sample_entry;
 mod codec;
 pub mod demux;
-pub mod demux_file_fmp4;
-pub mod demux_fmp4_segment;
+mod demux_fmp4_file;
+mod demux_fmp4_segment;
+mod demux_mp4_file;
 pub mod descriptors;
 pub mod mux;
 pub mod mux_fmp4_segment;
